pyproject.toml
[project]
name = "websocket-relay"
version = "0.1.0"
description = "Simple websocket-based relay server."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi",
    "uvicorn",
    "websockets",
    "python-dotenv",
    "aiohttp",
    "itsdangerous",
    "argon2-cffi",
]
